candlestick chart

+++ app/src/main/java/com/example/psx/domain/model/KLineModel.kt
@@ -18,6 +18,6 @@ data class KLineModelData (
     val open: Double,
     val high: Double,
     val low: Double,
-    val close: Long,
+    val close: Double,
     val volume: Long
 )
